Move statistical factors
