+++ zipline/pipeline/factors/factor.py
@@ -6,11 +6,6 @@ from operator import attrgetter
 from numbers import Number
 
 from numpy import inf, where
-from scipy.stats import (
-    linregress,
-    pearsonr,
-    spearmanr,
-)
 
 from zipline.errors import UnknownRankMethod
 from zipline.lib.normalize import naive_grouped_rowwise_apply
@@ -628,21 +623,20 @@ class Factor(RestrictedDTypeMixin, ComputableTerm):
         return Rank(self, method=method, ascending=ascending, mask=mask)
 
     @expect_types(
-        target_slice=Slice,
-        correlation_length=int,
-        mask=(Filter, NotSpecifiedType),
+        target=Slice, correlation_length=int, mask=(Filter, NotSpecifiedType),
     )
-    def rolling_pearsonr(self,
-                         target_slice,
-                         correlation_length,
-                         mask=NotSpecified):
+    def pearsonr(self, target, correlation_length, mask=NotSpecified):
         """
         Construct a new Factor that computes rolling pearson correlation
-        coefficients between the columns of `self` and the target slice.
+        coefficients between `target` and the columns of `self`.
+
+        This method can only be called on factors which are deemed safe for use
+        as inputs to other factors. This includes `Returns` and any factors
+        created from `Factor.rank` or `Factor.zscore`.
 
         Parameters
         ----------
-        target_slice : zipline.pipeline.slice.Slice
+        target : zipline.pipeline.slice.Slice
             The column of data with which to compute correlations against each
             column of data produced by `self`.
         correlation_length : int
@@ -655,32 +649,37 @@ class Factor(RestrictedDTypeMixin, ComputableTerm):
         Returns
         -------
         correlations : zipline.pipeline.factors.RollingPearson
-            A new Factor that will compute correlations between `target_slice`
-            and the columns of `self`.
+            A new Factor that will compute correlations between `target` and
+            the columns of `self`.
+
+        See Also
+        --------
+        :func:`scipy.stats.pearsonr`
+        :meth:`Factor.spearmanr`
         """
+        from .statistical import RollingPearson
         return RollingPearson(
             target_factor=self,
-            target_slice=target_slice,
+            target_slice=target,
             correlation_length=correlation_length,
             mask=mask,
         )
 
     @expect_types(
-        target_slice=Slice,
-        correlation_length=int,
-        mask=(Filter, NotSpecifiedType),
+        target=Slice, correlation_length=int, mask=(Filter, NotSpecifiedType),
     )
-    def rolling_spearmanr(self,
-                          target_slice,
-                          correlation_length,
-                          mask=NotSpecified):
+    def spearmanr(self, target, correlation_length, mask=NotSpecified):
         """
-        Construct a new Factor that computes rolling spearman correlation
-        coefficients between the columns of `self` and the target slice.
+        Construct a new Factor that computes rolling spearman rank correlation
+        coefficients between `target` and the columns of `self`.
+
+        This method can only be called on factors which are deemed safe for use
+        as inputs to other factors. This includes `Returns` and any factors
+        created from `Factor.rank` or `Factor.zscore`.
 
         Parameters
         ----------
-        target_slice : zipline.pipeline.slice.Slice
+        target : zipline.pipeline.slice.Slice
             The column of data with which to compute correlations against each
             column of data produced by `self`.
         correlation_length : int
@@ -693,32 +692,37 @@ class Factor(RestrictedDTypeMixin, ComputableTerm):
         Returns
         -------
         correlations : zipline.pipeline.factors.RollingSpearman
-            A new Factor that will compute correlations between `target_slice`
-            and the columns of `self`.
+            A new Factor that will compute correlations between `target` and
+            the columns of `self`.
+
+        See Also
+        --------
+        :func:`scipy.stats.spearmanr`
+        :meth:`Factor.pearsonr`
         """
+        from .statistical import RollingSpearman
         return RollingSpearman(
             target_factor=self,
-            target_slice=target_slice,
+            target_slice=target,
             correlation_length=correlation_length,
             mask=mask,
         )
 
     @expect_types(
-        target_slice=Slice,
-        regression_length=int,
-        mask=(Filter, NotSpecifiedType),
+        target=Slice, regression_length=int, mask=(Filter, NotSpecifiedType),
     )
-    def rolling_linear_regression(self,
-                                  target_slice,
-                                  regression_length,
-                                  mask=NotSpecified):
+    def linear_regression(self, target, regression_length, mask=NotSpecified):
         """
         Construct a new Factor that performs an ordinary least-squares
-        regression predicting the columns of `self` from `target_slice`.
+        regression predicting the columns of `self` from `target`.
+
+        This method can only be called on factors which are deemed safe for use
+        as inputs to other factors. This includes `Returns` and any factors
+        created from `Factor.rank` or `Factor.zscore`.
 
         Parameters
         ----------
-        target_slice : zipline.pipeline.slice.Slice
+        target : zipline.pipeline.slice.Slice
             The column of data to use as the predictor/independent variable in
             each regression.
         correlation_length : int
@@ -731,12 +735,17 @@ class Factor(RestrictedDTypeMixin, ComputableTerm):
         Returns
         -------
         regressions : zipline.pipeline.factors.RollingLinearRegression
-            A new Factor that will compute linear regressions of `target_slice`
+            A new Factor that will compute linear regressions of `target`
             against the columns of `self`.
+
+        See Also
+        --------
+        :func:`scipy.stats.linregress`
         """
+        from .statistical import RollingLinearRegression
         return RollingLinearRegression(
             target_factor=self,
-            target_slice=target_slice,
+            target_slice=target,
             regression_length=regression_length,
             mask=mask,
         )
@@ -1393,157 +1402,3 @@ class Latest(LatestMixin, CustomFactor):
 
     def compute(self, today, assets, out, data):
         out[:] = data[-1]
-
-
-class _RollingCorrelation(CustomFactor, SingleInputMixin):
-
-    def __new__(cls,
-                target_factor,
-                target_slice,
-                correlation_length,
-                mask=NotSpecified):
-        return super(_RollingCorrelation, cls).__new__(
-            cls,
-            inputs=[target_factor, target_slice],
-            window_length=correlation_length,
-            mask=mask,
-        )
-
-
-class RollingPearson(_RollingCorrelation):
-    """
-    A Factor that computes pearson correlation coefficients between a single
-    column of data and the columns of another Factor.
-
-    Parameters
-    ----------
-    target_factor : zipline.pipeline.factors.Factor
-        The factor for which to compute correlations of each of its columns
-        with `target_slice`.
-    target_slice : zipline.pipeline.slice.Slice
-        The column of data with which to compute correlations against each
-        column of data produced by `target_factor`.
-    correlation_length : int
-        Length of the lookback window over which to compute each
-        correlation coefficient.
-    mask : zipline.pipeline.Filter, optional
-        A Filter describing which assets (columns) of `target_factor` should
-        have their correlation with `target_slice` computed each day.
-
-    See Also
-    --------
-    :func:`scipy.stats.pearsonr`
-    :class:`Factor.rolling_pearsonr`
-    :class:`zipline.pipeline.factors.RollingPearsonOfReturns`
-
-    Notes
-    -----
-    Most users should call Factor.rolling_pearsonr rather than directly
-    construct an instance of this class.
-    """
-    def compute(self, today, assets, out, factor_data, slice_data):
-        slice_data_column = slice_data[:, 0]
-        for i in range(len(out)):
-            # pearsonr returns the R-value and the P-value.
-            out[i] = pearsonr(factor_data[:, i], slice_data_column)[0]
-
-
-class RollingSpearman(_RollingCorrelation):
-    """
-    A Factor that computes spearman rank correlation coefficients between a
-    single column of data and the columns of another Factor.
-
-    Parameters
-    ----------
-    target_factor : zipline.pipeline.factors.Factor
-        The factor for which to compute correlations of each of its columns
-        with `target_slice`.
-    target_slice : zipline.pipeline.slice.Slice
-        The column of data with which to compute correlations against each
-        column of data produced by `target_factor`.
-    correlation_length : int
-        Length of the lookback window over which to compute each
-        correlation coefficient.
-    mask : zipline.pipeline.Filter, optional
-        A Filter describing which assets (columns) of `target_factor` should
-        have their correlation with `target_slice` computed each day.
-
-    See Also
-    --------
-    :func:`scipy.stats.spearmanr`
-    :class:`Factor.rolling_spearmanr`
-    :class:`zipline.pipeline.factors.RollingSpearmanOfReturns`
-
-    Notes
-    -----
-    Most users should call Factor.rolling_spearmanr rather than directly
-    construct an instance of this class.
-    """
-    def compute(self, today, assets, out, factor_data, slice_data):
-        slice_data_column = slice_data[:, 0]
-        for i in range(len(out)):
-            # spearmanr returns the R-value and the P-value.
-            out[i] = spearmanr(factor_data[:, i], slice_data_column)[0]
-
-
-class RollingLinearRegression(CustomFactor, SingleInputMixin):
-    """
-    A Factor that performs an ordinary least-squares regression predicting the
-    columns of another Factor from a single column of data.
-
-    Parameters
-    ----------
-    target_factor : zipline.pipeline.factors.Factor
-        The factor whose columns are the predicted/dependent variable of each
-        regression with `target_slice`.
-    target_slice : zipline.pipeline.slice.Slice
-        The column of data to use as the predictor/independent variable in
-        each regression with the columns of `target_factor`.
-    correlation_length : int
-        Length of the lookback window over which to compute each regression.
-    mask : zipline.pipeline.Filter, optional
-        A Filter describing which assets (columns) of `target_factor` should be
-        regressed against `target_slice` each day.
-
-    See Also
-    --------
-    :func:`scipy.stats.linregress`
-    :class:`Factor.rolling_linear_regression`
-    :class:`zipline.pipeline.factors.RollingLinearRegressionOfReturns`
-
-    Notes
-    -----
-    Most users should call Factor.rolling_linear_regression rather than
-    directly construct an instance of this class.
-    """
-    outputs = ['alpha', 'beta', 'r_value', 'p_value', 'stderr']
-
-    def __new__(cls,
-                target_factor,
-                target_slice,
-                regression_length,
-                mask=NotSpecified):
-        return super(RollingLinearRegression, cls).__new__(
-            cls,
-            inputs=[target_factor, target_slice],
-            window_length=regression_length,
-            mask=mask,
-        )
-
-    def compute(self, today, assets, out, factor_data, slice_data):
-        slice_data_column = slice_data[:, 0]
-
-        alpha = out.alpha
-        beta = out.beta
-        r_value = out.r_value
-        p_value = out.p_value
-        stderr = out.stderr
-        for i in range(len(out)):
-            regr_results = linregress(y=factor_data[:, i], x=slice_data_column)
-            # `linregress` returns its results in the following order:
-            # slope, intercept, r-value, p-value, stderr
-            alpha[i] = regr_results[1]
-            beta[i] = regr_results[0]
-            r_value[i] = regr_results[2]
-            p_value[i] = regr_results[3]
-            stderr[i] = regr_results[4]
