+++ zipline/pipeline/factors/statistical.py
@@ -1,13 +1,170 @@
 
+from scipy.stats import (
+    linregress,
+    pearsonr,
+    spearmanr,
+)
+
+from zipline.pipeline.factors import CustomFactor
 from zipline.pipeline.filters import SingleAsset
+from zipline.pipeline.mixins import SingleInputMixin
 from zipline.pipeline.term import AssetExists, NotSpecified
 
-from .factor import (
-    RollingLinearRegression,
-    RollingPearson,
-    RollingSpearman,
-)
 from .technical import Returns
+
+
+class _RollingCorrelation(CustomFactor, SingleInputMixin):
+
+    def __new__(cls,
+                target_factor,
+                target_slice,
+                correlation_length,
+                mask=NotSpecified):
+        return super(_RollingCorrelation, cls).__new__(
+            cls,
+            inputs=[target_factor, target_slice],
+            window_length=correlation_length,
+            mask=mask,
+        )
+
+
+class RollingPearson(_RollingCorrelation):
+    """
+    A Factor that computes pearson correlation coefficients between a single
+    column of data and the columns of another Factor.
+
+    Parameters
+    ----------
+    target_factor : zipline.pipeline.factors.Factor
+        The factor for which to compute correlations of each of its columns
+        with `target_slice`.
+    target_slice : zipline.pipeline.slice.Slice
+        The column of data with which to compute correlations against each
+        column of data produced by `target_factor`.
+    correlation_length : int
+        Length of the lookback window over which to compute each
+        correlation coefficient.
+    mask : zipline.pipeline.Filter, optional
+        A Filter describing which assets (columns) of `target_factor` should
+        have their correlation with `target_slice` computed each day.
+
+    See Also
+    --------
+    :func:`scipy.stats.pearsonr`
+    :meth:`Factor.pearsonr`
+    :class:`zipline.pipeline.factors.RollingPearsonOfReturns`
+
+    Notes
+    -----
+    Most users should call Factor.pearsonr rather than directly construct an
+    instance of this class.
+    """
+    def compute(self, today, assets, out, factor_data, slice_data):
+        slice_data_column = slice_data[:, 0]
+        for i in range(len(out)):
+            # pearsonr returns the R-value and the P-value.
+            out[i] = pearsonr(factor_data[:, i], slice_data_column)[0]
+
+
+class RollingSpearman(_RollingCorrelation):
+    """
+    A Factor that computes spearman rank correlation coefficients between a
+    single column of data and the columns of another Factor.
+
+    Parameters
+    ----------
+    target_factor : zipline.pipeline.factors.Factor
+        The factor for which to compute correlations of each of its columns
+        with `target_slice`.
+    target_slice : zipline.pipeline.slice.Slice
+        The column of data with which to compute correlations against each
+        column of data produced by `target_factor`.
+    correlation_length : int
+        Length of the lookback window over which to compute each
+        correlation coefficient.
+    mask : zipline.pipeline.Filter, optional
+        A Filter describing which assets (columns) of `target_factor` should
+        have their correlation with `target_slice` computed each day.
+
+    See Also
+    --------
+    :func:`scipy.stats.spearmanr`
+    :meth:`Factor.spearmanr`
+    :class:`zipline.pipeline.factors.RollingSpearmanOfReturns`
+
+    Notes
+    -----
+    Most users should call Factor.spearmanr rather than directly construct an
+    instance of this class.
+    """
+    def compute(self, today, assets, out, factor_data, slice_data):
+        slice_data_column = slice_data[:, 0]
+        for i in range(len(out)):
+            # spearmanr returns the R-value and the P-value.
+            out[i] = spearmanr(factor_data[:, i], slice_data_column)[0]
+
+
+class RollingLinearRegression(CustomFactor, SingleInputMixin):
+    """
+    A Factor that performs an ordinary least-squares regression predicting the
+    columns of another Factor from a single column of data.
+
+    Parameters
+    ----------
+    target_factor : zipline.pipeline.factors.Factor
+        The factor whose columns are the predicted/dependent variable of each
+        regression with `target_slice`.
+    target_slice : zipline.pipeline.slice.Slice
+        The column of data to use as the predictor/independent variable in
+        each regression with the columns of `target_factor`.
+    correlation_length : int
+        Length of the lookback window over which to compute each regression.
+    mask : zipline.pipeline.Filter, optional
+        A Filter describing which assets (columns) of `target_factor` should be
+        regressed against `target_slice` each day.
+
+    See Also
+    --------
+    :func:`scipy.stats.linregress`
+    :meth:`Factor.linear_regression`
+    :class:`zipline.pipeline.factors.RollingLinearRegressionOfReturns`
+
+    Notes
+    -----
+    Most users should call Factor.linear_regression rather than directly
+    construct an instance of this class.
+    """
+    outputs = ['alpha', 'beta', 'r_value', 'p_value', 'stderr']
+
+    def __new__(cls,
+                target_factor,
+                target_slice,
+                regression_length,
+                mask=NotSpecified):
+        return super(RollingLinearRegression, cls).__new__(
+            cls,
+            inputs=[target_factor, target_slice],
+            window_length=regression_length,
+            mask=mask,
+        )
+
+    def compute(self, today, assets, out, factor_data, slice_data):
+        slice_data_column = slice_data[:, 0]
+
+        alpha = out.alpha
+        beta = out.beta
+        r_value = out.r_value
+        p_value = out.p_value
+        stderr = out.stderr
+        for i in range(len(out)):
+            regr_results = linregress(y=factor_data[:, i], x=slice_data_column)
+            # `linregress` returns its results in the following order:
+            # slope, intercept, r-value, p-value, stderr
+            alpha[i] = regr_results[1]
+            beta[i] = regr_results[0]
+            r_value[i] = regr_results[2]
+            p_value[i] = regr_results[3]
+            stderr[i] = regr_results[4]
 
 
 class RollingPearsonOfReturns(RollingPearson):
